==> Tools/Step1_Build_Perception_models.Rmd <==
---
title: "R Notebook"
output:
html_notebook: default
---
<!-- For a given set of drugs of Interest - Build a PERCEPTION model -->
```{r, setup, include=FALSE}
# Set working directory to the directory where "PERCEPTION" is cloned.
working_directory='/Users/sinhas8/PERCEPTION/'
knitr::opts_knit$set(root.dir = working_directory)
```
<!-- Load DepMAP data and already existing models -->
```{r}
onTarget=readRDS('Data/DepMapv12.RDS')
# Function to load a RMD file to another
ksource <- function(x, ...) {
  library(knitr)
  source(purl(x, output = tempfile()), ...)
}
ksource('Tools/step0A_functions_needed.Rmd')
ksource('Tools/step0B_functions_needed.Rmd')
# <!-- Load already built models for all FDA approved drugs -->
approved_drugs_model=readRDS('Data/FDA_approved_drugs_models.RDS')
```
<!-- Build PERCEPTION Models -->
<!-- Step 0: Create Inputs for PERCEPTION -->
```{r}
# DOI stands for Drugs of interest; PERCEPTION will be run on these drugs; For each drug, there will be a seperate model
DOI=names(approved_drugs_model)[1:44]
#@Sumit - The above run is for two drugs. It takes 1 minute or so to run. If you use all 44 - this may take 15-20 mins. 
# What cancer type the PERCEPTION will be trained on?; PanCan stands for Pan-cancer.
cancerType='PanCan'
# What features are we using?
genesUsed_toBuild=intersect(rownames(onTarget$expression_20Q4),
                            rownames(onTarget$scRNA_complete))
# An expression-normalization step is needed to be run before we run PERCEPITON.
# How do we do that?
# onTarget$scRNA_subset_rnorm=rank_normalization_mat(onTarget$scRNA_complete[genesUsed_toBuild,])
# onTarget$expression_rnorm=rank_normalization_mat(onTarget$expression_20Q4[genesUsed_toBuild,])

# How many features should the model use?
Possible_Number_of_features=nrow(onTarget$expression_rnorm) * seq(0.0005, 0.02, length = 5)
```
<!-- Step 1 feature Selection/Ranking using Bulk -->
```{r message=FALSE, warning=FALSE, paged.print=FALSE}
features_bulk_all_cancer_type_Cavg=run_parallel_feature_ranking_bulk(id_cancerType = cancerType,
                                                                     infunc_DrugsToUse=DOI,
                                                                     resp_measure_mode = 'AUC',
                                                                     infunc_exclude_cancer = 'PanCan')
#How feature ranking looks like?
head(features_bulk_all_cancer_type_Cavg[[1]], 5)
```
<!-- Initialize list to save perception output; Provide cores for parallelization -->
```{r message=FALSE, warning=FALSE, paged.print=FALSE}
Raw_models_output <- list();
Performance_by_features <- list(); for_output_lung_Test_vglm <- list()
cl <- parallel::makeForkCluster(4)
doParallel::registerDoParallel(cl)
```
<!-- Step 2-3: Build and tune PERCEPTION -->
<!-- In the below code, the outer loop iterates over the drugs of interest (DOI) and the inner loop iterates over a parameter of input genes. The build_on_BULK_v2 function is called to build the model, and the results are stored in the Raw_models_output list. The performance of each model is printed and stored in the Performance_by_features list. The best model for each drug is stored in the for_output_lung_Test_vglm list. The final models are saved to a file. -->
```{r message=FALSE, warning=FALSE, paged.print=FALSE}
# Iterate over the DOI (drugs of interest)
for (i in 1:length(DOI)) {
  # Initialize variables
  Raw_models_output <- list()
  id_counter <- 1
  # For each given drug; iterate over a parameter of 'features count'
  for (infunc_k_features_grid in Possible_Number_of_features ) {
    # Build the model
    Raw_models_output[[id_counter]] <- build_on_BULK_v2(
      infunc_drugName = DOI[i],
      infunc_cancerType = cancerType,
      infunc_features = rownames(features_bulk_all_cancer_type_Cavg[[i]]),
      single_best = rownames(features_bulk_all_cancer_type_Cavg[[i]])[1],
      k_features = infunc_k_features_grid,
      mode = 'AUC',
      model_type = 'glmnet',
      exclude_cancer = 'PanCan'
    )
    id_counter <- id_counter + 1
  }

  # Print and store the performance
  print(sapply(Raw_models_output, function(x) x$performance_in_scRNA))
  Performance_by_features[[i]] <- lapply(Raw_models_output, function(x) x$performance_in_scRNA)

  # Store the Tuned model
  for_output_lung_Test_vglm[[i]] <- err_handle(Raw_models_output[[which.max(sapply(Raw_models_output, function(x) x$performance_in_scRNA)[2,])]])
  
  # Remove temporary variables and garbage collect
  rm(Raw_models_output)
  gc()
}

# Stop parallel processing
parallel::stopCluster(cl)

# Assign names to the TUned models and save them
names(for_output_lung_Test_vglm) <- DOI
Tuned_models_output <- for_output_lung_Test_vglm
saveRDS(Tuned_models_output, paste('Data/PERCEPTION_models_pancan', Sys.time(), '.RDS',sep=''))
```
