==> Tools/Step2_Figure1.Rmd <==
---
title: "R Notebook"
output: html_notebook
---

This is an [R Markdown](http://rmarkdown.rstudio.com) Notebook. When you execute code within the notebook, the results appear beneath the code. 

Try executing this chunk by clicking the *Run* button within the chunk or by placing your cursor inside it and pressing *Cmd+Shift+Enter*. 
```{r, setup, include=FALSE}
# Set working directory to the directory where "PERCEPTION" is cloned.
working_directory='/Users/sinhas8/PERCEPTION/'
knitr::opts_knit$set(root.dir = working_directory)
```
<!-- Plot Figure 1D-E:  -->
<!-- Plot 1D-E represents single cells from cell lines (lung or non-lung) from DepMap-->
```{r}
onTarget=readRDS('Data/DepMapv12.RDS')
# Function to load a RMD file to another
ksource <- function(x, ...) {
  library(knitr)
  source(purl(x, output = tempfile()), ...)
}
ksource('Tools/step0A_functions_needed.Rmd')
ksource('Tools/step0B_functions_needed.Rmd')
# <!-- Load already built models for all FDA approved drugs -->
approved_drugs_model=readRDS('Data/FDA_approved_drugs_models.RDS')
```
<!-- Load required TSNE mapping information for each single cells -->
```{r}
# Co-ordinates of t-SNE plot of single from Kinker et al. from the link below
# https://singlecell.broadinstitute.org/single_cell/study/SCP542/pan-cancer-cell-line-heterogeneity#study-download
lung_tSNE=read.csv('Data/lung_tSNE.txt', sep='\t')
lung_tSNE=lung_tSNE[-1,]
lung_tSNE[,2]=as.numeric(lung_tSNE[,2])
lung_tSNE[,3]=as.numeric(lung_tSNE[,3])
```
<!-- Biomarker Preprocessing -->
<!-- TSNE plots -->
```{r}
biomarker_name='MDM2'
biomarker_id=match(biomarker_name, rownames(onTarget$CPM_scRNA_CCLE_rnorm))
biomarker_abundance=onTarget$CPM_scRNA_CCLE_rnorm[biomarker_id,]
needed_cells_id=match(lung_tSNE$NAME, colnames(onTarget$CPM_scRNA_CCLE_rnorm))
matched_biomarker_abundance=biomarker_abundance[needed_cells_id]
# Predict viability
mdm2_model=readRDS('Data/Response_model_nutlin3.RDS')
viability_required=viability_from_model(infunc_DruName=DOI,
                                        infun_model=mdm2_model$nutlin3$model,
                                        infunc_dataset=onTarget$CPM_scRNA_CCLE_rnorm[
                                        match(mdm2_model$nutlin3$model$coefnames, 
                                                rownames(onTarget$CPM_scRNA_CCLE_rnorm)),
                                          needed_cells_id])
lung_tSNE$biomarker=range01(matched_biomarker_abundance)
lung_tSNE$biomarker_scaled=range01(rank(matched_biomarker_abundance))
lung_tSNE$killing=range01(-viability_required)
lung_tSNE$killing_scaled=range01(rank(-viability_required))
# Biomarker Expression
lung_tSNE$DepMapID=onTarget$metadata_CPM_scRNA$DepMap_ID[match(lung_tSNE$NAME, onTarget$metadata_CPM_scRNA$NAME)]
lung_tSNE$p53=onTarget$mutations_matrix['TP53',match(lung_tSNE$DepMapID, colnames(onTarget$mutations_matrix))]
correalation_by_p53=sapply(split(lung_tSNE, lung_tSNE$p53), function(x)
  unlist(cor.test_trimmed_v0(x$biomarker, x$killing)))
```
<!-- MDM2 expression in P53 WT cells -->
```{r}
WT_p53=which(lung_tSNE$p53==0)
length(lung_tSNE$DepMapID)
cor.test_trimmed_v0(lung_tSNE$biomarker[WT_p53], lung_tSNE$killing[WT_p53])
plot_tsne_biomarker <- ggplot(lung_tSNE[WT_p53,], aes(x=X, y=Y, color=biomarker_scaled))+
  geom_point(size=0.2)+
  theme_bw(base_size = 8)+
  labs(color='MDM2  Exp')+
  theme(legend.position='top')+
  # scale_color_viridis_c()
  scale_colour_gradientn(colours = c("#F8766D",'lightgrey' ,"#00BFC4"))+
  labs(x='', y='')
# MDM2 Killing

plot_tsne_Killing <- ggplot(lung_tSNE[WT_p53,], aes(x=X, y=Y, color=killing_scaled))+
  geom_point(size=0.2)+
  theme_bw(base_size = 8)+
  labs(color='nutlin-3 Killing')+
  theme(legend.position='top')+
  scale_colour_gradientn(colours = c("#F8766D",'lightgrey' ,"#00BFC4"))+
  labs(x='', y='')
tsne_together <- grid.arrange(plot_tsne_biomarker, plot_tsne_Killing, nrow=1)
ggsave(tsne_together, filename = 'Figures/Figure1D_nutlin_v1.pdf', height=2.5, width=5.6)
ggsave(tsne_together, filename = 'Figures/Figure1D_nutlin_v1.png', height=2.5, width=5.6)
```
<!-- Plotting Figure 1E: Erlotinib vs EGFR pathway activity Figure -->
<!-- Load The models to Predict Killing -->
```{r}
needed_cells_id=match(lung_tSNE$NAME, colnames(onTarget$CPM_scRNA_CCLE_rnorm))
# Predict viability
killing_matrix_list=lapply(1:length(approved_drugs_model), function(x) 
  viability_required=viability_from_model(
    infunc_DruName=names(approved_drugs_model)[x],
    infun_model=approved_drugs_model[[x]]$model,
    infunc_dataset=onTarget$CPM_scRNA_CCLE_rnorm[
      match(approved_drugs_model[[x]]$model$coefnames, 
            rownames(onTarget$CPM_scRNA_CCLE_rnorm)),
      needed_cells_id] ))
names(killing_matrix_list)=names(approved_drugs_model)
killing_matrix_list=do.call(cbind, killing_matrix_list)
dim(killing_matrix_list)
expanded_lung_tSNE=cbind(lung_tSNE, killing_matrix_list,
                         onTarget$metadata_CPM_scRNA[
                           match(lung_tSNE$NAME,onTarget$metadata_CPM_scRNA$NAME), 10:21])
```
<!-- Plot TSNE -->
```{r}
expanded_lung_tSNE$Y=as.numeric(expanded_lung_tSNE$Y)
expanded_lung_tSNE$Cell_line=onTarget$metadata_CPM_scRNA$DepMap_ID[match(expanded_lung_tSNE$NAME, onTarget$metadata_CPM_scRNA$NAME)]
expanded_lung_tSNE$EGFR_mutation=onTarget$mutations_matrix['EGFR',
                          match(expanded_lung_tSNE$Cell_line, colnames(onTarget$mutations_matrix))]
EGFR_WT_signature=read.csv('Data/EGFR_WT_signature.csv')
# EGFR pathway activity
egfr_mut_signature_activity=colMeans(onTarget$CPM_scRNA_CCLE_rnorm[EGFR_WT_signature$x,])
expanded_lung_tSNE$egfr_mut_signature_activity=egfr_mut_signature_activity[match(expanded_lung_tSNE$NAME, names(egfr_mut_signature_activity))]
plot_tsne_biomarker <- ggplot(expanded_lung_tSNE, aes(x=X, y=Y,
                                                      color= range01(rank(-as.numeric(egfr_mut_signature_activity)))
                                                      ))+
  geom_point(size=0.5)+
  theme_bw(base_size = 8)+
  labs(color='EGFR activity')+
  theme(legend.position='top')+
  # scale_color_viridis_c()
  scale_colour_gradientn(colours = c("#F8766D",'lightgrey' ,"#00BFC4"))+
  labs(x='', y='')

plot_tsne_Killing <- ggplot(expanded_lung_tSNE, aes(x=X, y=Y,
                                                    color=range01(rank(-as.numeric(erlotinib)))
                                                    ))+
  geom_point(size=0.5)+
  theme_bw(base_size = 8)+
  labs(color='gefitinib Killing')+
  theme(legend.position='top')+
  scale_colour_gradientn(colours = c("#F8766D",'lightgrey' ,"#00BFC4"))+
  labs(x='', y='')

tsne_together_gefitinib <- grid.arrange(plot_tsne_biomarker, plot_tsne_Killing, nrow=1)
ggsave(tsne_together_gefitinib, filename = 'Figures/Figure1_erlotinibv1.pdf', height=2.5, width=5.6)
```
<!-- Figure 1B-C -->
<!-- Performance of our models in cross validation -->
```{r}
performance_extended_grid=readRDS('Data/model_performances.RDS')
drugNames_fda_approved=names(performance_extended_grid)
predictive_drugs=drugNames_fda_approved[which(sapply(performance_extended_grid, function(x) x$performance_in_scRNA[2]>0.3))]
model_perf=performance_extended_grid
```
<!-- Panel B Plot performance in scRNA-seq -->
```{r}
performance_in_scRNA=data.frame(do.call(rbind, lapply(model_perf, function(x) x$performance_in_scRNA)))
performance_in_bulk=data.frame(do.call(rbind, lapply(model_perf, function(x) x$performance_in_bulk)))
performance_in_pseudo_bulk=data.frame(do.call(rbind, lapply(model_perf, function(x) x$performance_in_pseudo_bulk)))
df2plot=rbind(data.frame(drugsCount=sapply(seq(0.1, 0.6, 0.01), function(x) sum(performance_in_scRNA$estimate.cor>x)),
                 dataused='scRNA-seq',
                 Predictibility=seq(0.1, 0.6, 0.01)),
      data.frame(drugsCount=sapply(seq(0.1, 0.6, 0.01), function(x) sum(performance_in_bulk$estimate.cor>x)),
                 dataused='bulk',
                 Predictibility=seq(0.1, 0.6, 0.01)),
      data.frame(drugsCount=sapply(seq(0.1, 0.6, 0.01), function(x) sum(performance_in_pseudo_bulk$estimate.cor>x)),
                 dataused='psuedo-bulk',
                 Predictibility=seq(0.1, 0.6, 0.01))
)
PanelB <- ggplot(df2plot, aes(x=Predictibility, y=drugsCount, color=dataused))+
  geom_point()+
  geom_line()+
  geom_vline(xintercept = 0.3, linetype='dashed')+
  theme_bw(base_size = 20)+
  labs(y='Number of Drugs', color='Dataset Used', x='Predictibility\n (Pearson Correlation)')+
  theme(legend.position = 'none')
```

```{r}
performance_in_scRNA$dataused='scRNA'
performance_in_bulk$dataused='bulk'
performance_in_pseudo_bulk$dataused='pseudo-bulk'
df2plot=rbind(performance_in_scRNA, performance_in_bulk, performance_in_pseudo_bulk)
df2plot$drugName=drugNames_fda_approved
df2plot$drug_category=factor(onTarget$drugCategory$drug_category[match(df2plot$drugName,
                                                              stripall2match(onTarget$drugCategory$name))],
                             labels = c('Chemo', 'Targeted'))
PanelC <- ggplot(df2plot[df2plot$estimate.cor>0.3,], aes(x=drug_category,y=estimate.cor, color=dataused))+
  geom_boxplot()+
  theme_bw(base_size = 20)+
  labs(y='Pearson Correlation', color='Dataset Used', x='Drug Category')+
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust=1))
```
<!-- Add both tSNE and Figure BC -->
```{r}
tsne_together <- grid.arrange(plot_tsne_biomarker, plot_tsne_Killing, nrow=1)
Figure1_BC=grid.arrange(PanelB, PanelC, nrow = 1)
ggsave(grid.arrange(Figure1_BC, tsne_together, nrow=2),
       filename = 'Figures/Figure1BCDE.pdf', width=12, height=12)
```

Add a new chunk by clicking the *Insert Chunk* button on the toolbar or by pressing *Cmd+Option+I*.

When you save the notebook, an HTML file containing the code and output will be saved alongside it (click the *Preview* button or press *Cmd+Shift+K* to preview the HTML file). 

The preview shows you a rendered HTML copy of the contents of the editor. Consequently, unlike *Knit*, *Preview* does not run any R code chunks. Instead, the output of the chunk when it was last run in the editor is displayed
